refactor: 심야자습 출석 API 구조 개선
