+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/presentation/nightstudy/NightStudyController.kt
@@ -8,20 +8,15 @@ import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.Assign
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.PersonalNightStudyApplyRequest
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.ProjectNightStudyApplyRequest
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.RejectNightStudyRequest
-import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.request.UpdateNightStudyAttendanceRequest
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.NightStudyApplicationResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.NightStudyApprovedCountResponse
-import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.NightStudyAttendanceCountResponse
-import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.NightStudyAttendanceResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.PersonalNightStudyResponse
 import com.b1nd.dodamdodam.nightstudy.application.nightstudy.data.response.ProjectNightStudyResponse
 import com.b1nd.dodamdodam.nightstudy.domain.nightstudy.enumeration.NightStudyStatusType
 import com.b1nd.dodamdodam.nightstudy.domain.nightstudy.enumeration.NightStudyType
 import com.b1nd.dodamdodam.core.common.data.InfinityScrollPageResponse
 import org.springframework.data.domain.Pageable
-import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.web.bind.annotation.*
-import java.time.LocalDate
 import java.util.UUID
 
 @RestController
@@ -92,31 +87,4 @@ class NightStudyController(
     @DeleteMapping("/applications/{id}/room")
     fun unassignRoom(@PathVariable id: UUID): Response<Any> =
         nightStudyUseCase.unassignRoom(id)
-
-    @UserAccess(roles = [RoleType.DORMITORY_MANAGER])
-    @GetMapping("/attendance/count")
-    fun countAttendance(
-        @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd") date: LocalDate?,
-        @RequestParam period: Int,
-    ): Response<NightStudyAttendanceCountResponse> =
-        nightStudyUseCase.countAttendance(date ?: LocalDate.now(), period)
-
-    @UserAccess(roles = [RoleType.DORMITORY_MANAGER])
-    @PatchMapping("/attendance/{userId}")
-    fun updateAttendance(
-        @PathVariable userId: UUID,
-        @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd") date: LocalDate?,
-        @RequestParam period: Int,
-        @RequestBody request: UpdateNightStudyAttendanceRequest
-    ): Response<NightStudyAttendanceResponse> =
-        nightStudyUseCase.updateAttendance(userId, date ?: LocalDate.now(), period, request)
-
-    @UserAccess(roles = [RoleType.DORMITORY_MANAGER])
-    @GetMapping("/attendance/{userId}")
-    fun getAttendance(
-        @PathVariable userId: UUID,
-        @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd") date: LocalDate?,
-        @RequestParam period: Int,
-    ): Response<NightStudyAttendanceResponse> =
-        nightStudyUseCase.getAttendance(userId, date ?: LocalDate.now(), period)
 }
